Flip segment from iterating up to iterating down
The reason for this is the following:
When iterating up, one starts at 0 and goes up to how many segments
need to be sent until the full array of information is fully
transmitted.

The problem is on the receiving end, one cannot just call Receive()
at the Transport layer, since one does not know the total number of
segments. So this information either needs to somehow be transmitted,
or a delegation to ReceiveNumberOfSegments needs to be asked from
Transport to Network, to Datalink and finally to Physical.

My iterating down, the receiver can continue calling the lower Netork
Receive() and decrementing the segment index until it reaches zero.